refactor(Plant): move out from PVZ.h
