+++ pvzclass/Classes/Plant.cpp
@@ -1,4 +1,3 @@
-#include "../PVZ.h"
 #include "GameObject.hpp"
 
 PVZ::Plant::Plant(int indexoraddress)
